requirements.txt
firebase_functions~=0.1.0

tensorflow>=2.0
matplotlib>=3.0
numpy>=1.16
Pillow>=7.0
six>=1.12